chore: pox3 to pox4 misc renames
