+++ .vscode/launch.json
@@ -316,14 +316,14 @@
     {
       "type": "node",
       "request": "launch",
-      "name": "Jest: 2.4",
+      "name": "Jest: 2.5",
       "program": "${workspaceFolder}/node_modules/.bin/jest",
       "args": [
         "--testTimeout=3600000",
         "--runInBand",
         "--no-cache",
         "--config",
-        "${workspaceRoot}/tests/jest.config.2.4.js"
+        "${workspaceRoot}/tests/jest.config.2.5.js",
       ],
       "outputCapture": "std",
       "console": "integratedTerminal",
