+++ src/tests-2.5/block-zero-handling.ts
