+++ src/tests-2.5/global-setup.ts
