+++ src/tests-2.5/pox-4-burnchain-delegate-stx.ts
@@ -172,7 +172,7 @@ async function createPox2DelegateStx(args: {
   };
 }
 
-describe('PoX-3 - Stack using Bitcoin-chain ops', () => {
+describe('PoX-4 - Stack using Bitcoin-chain ops', () => {
   const seedAccount = testnetKeys[0];
 
   let db: PgWriteStore;
