+++ src/tests-2.5/pox-4-rosetta-btc-addr-types.ts
@@ -21,7 +21,7 @@ const BTC_ADDRESS_CASES = [
 ] as const;
 
 describe.each(BTC_ADDRESS_CASES)(
-  'PoX-3 - Rosetta - Stack with BTC address format $addressFormat',
+  'PoX-4 - Rosetta - Stack with BTC address format $addressFormat',
   ({ addressFormat }) => {
     let poxInfo: CoreRpcPoxInfo;
     const account = testnetKeys[1];
